requirements.txt
redis
pysnmp
netaddr
click
pyyaml
paramiko
requests
